LPS-184427 Move PublishedAppsDashboardPage functions to a util file
